Adjust to a DI structure
