+++ gotrue-csharp.sln
@@ -1,18 +1,20 @@
 ﻿
 Microsoft Visual Studio Solution File, Format Version 12.00
-# Visual Studio Version 16
-VisualStudioVersion = 16.0.808.6
+# Visual Studio Version 17
+VisualStudioVersion = 17.2.32519.379
 MinimumVisualStudioVersion = 10.0.40219.1
-Project("{FAE04EC0-301F-11D3-BF4B-00C04F79EFBC}") = "Gotrue", "Gotrue\Gotrue.csproj", "{88818A9F-8595-481A-A8FB-039D86118CE5}"
+Project("{9A19103F-16F7-4668-BE54-9A1E7A4F7556}") = "Gotrue", "Gotrue\Gotrue.csproj", "{88818A9F-8595-481A-A8FB-039D86118CE5}"
 EndProject
-Project("{FAE04EC0-301F-11D3-BF4B-00C04F79EFBC}") = "GotrueTests", "GotrueTests\GotrueTests.csproj", "{B2C4F4BD-572E-4182-A2C9-8E5C9F6AD04C}"
+Project("{9A19103F-16F7-4668-BE54-9A1E7A4F7556}") = "GotrueTests", "GotrueTests\GotrueTests.csproj", "{B2C4F4BD-572E-4182-A2C9-8E5C9F6AD04C}"
 EndProject
 Project("{2150E333-8FDC-42A3-9474-1A3956D46DE8}") = "Solution Items", "Solution Items", "{228693BB-A395-4123-93E8-5299FF875615}"
 	ProjectSection(SolutionItems) = preProject
+		CHANGELOG.md = CHANGELOG.md
 		docker-compose.yml = docker-compose.yml
 		README.md = README.md
-		CHANGELOG.md = CHANGELOG.md
 	EndProjectSection
+EndProject
+Project("{FAE04EC0-301F-11D3-BF4B-00C04F79EFBC}") = "GotrueExample", "GotrueExample\GotrueExample.csproj", "{7DD97085-C72E-43B8-A55B-2DFBF6F1AB03}"
 EndProject
 Global
 	GlobalSection(SolutionConfigurationPlatforms) = preSolution
@@ -28,6 +30,10 @@ Global
 		{B2C4F4BD-572E-4182-A2C9-8E5C9F6AD04C}.Debug|Any CPU.Build.0 = Debug|Any CPU
 		{B2C4F4BD-572E-4182-A2C9-8E5C9F6AD04C}.Release|Any CPU.ActiveCfg = Release|Any CPU
 		{B2C4F4BD-572E-4182-A2C9-8E5C9F6AD04C}.Release|Any CPU.Build.0 = Release|Any CPU
+		{7DD97085-C72E-43B8-A55B-2DFBF6F1AB03}.Debug|Any CPU.ActiveCfg = Debug|Any CPU
+		{7DD97085-C72E-43B8-A55B-2DFBF6F1AB03}.Debug|Any CPU.Build.0 = Debug|Any CPU
+		{7DD97085-C72E-43B8-A55B-2DFBF6F1AB03}.Release|Any CPU.ActiveCfg = Release|Any CPU
+		{7DD97085-C72E-43B8-A55B-2DFBF6F1AB03}.Release|Any CPU.Build.0 = Release|Any CPU
 	EndGlobalSection
 	GlobalSection(SolutionProperties) = preSolution
 		HideSolutionNode = FALSE
